push commit

+++ fundamentals/18_promo/promotion1.java
@@ -2,7 +2,7 @@ public class Promote1{
     public static void main(String[] args) {
         byte a = 5;
         byte b = 2;
-        int c = a +b;     
+        byte c = (byte)(a +b);     
         System.out.println(c);
     }
 }
